requirements.txt
boto3==1.9.94
botocore==1.12.94
Django==2.1.7
django-jet==1.0.8
django-model-utils==3.1.2
django-storages==1.7.1
docutils==0.14
jmespath==0.9.3
Pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
djangorestframework==3.9.2
